refactor(test) : RacingGameTest 추가
